requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.29.0
uvloop==0.21.0
pydantic==2.7.0
pydantic[email]==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
loguru==0.7.2
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
watchtower==3.0.0
aws-xray-sdk==2.13.0
mangum==0.17.0
starlette==0.47.2
python-json-logger==2.0.7
jinja2==3.1.6
python-multipart==0.0.18
requests==2.32.4
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
